# apps/visualizer/backend-ts/tests/fixtures/flask-openapi.json
{
...
 "components": {
  "schemas": {
...
   "ErrorBody.45d9b59": {
    "properties": {
     "code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Code"
     },
     "error": {
      "title": "Error",
      "type": "string"
     }
    },
    "required": [
     "error"
    ],
    "title": "ErrorBody",
    "type": "object"
   },
...
   "PerspectiveDetail.15b0cf1": {
    "additionalProperties": false,
    "properties": {
     "active": {
      "title": "Active",
      "type": "boolean"
     },
     "created_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Created At"
     },
     "description": {
      "title": "Description",
      "type": "string"
     },
     "display_name": {
      "title": "Display Name",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "optional": {
      "title": "Optional",
      "type": "boolean"
     },
     "prompt_markdown": {
      "title": "Prompt Markdown",
      "type": "string"
     },
     "slug": {
      "title": "Slug",
      "type": "string"
     },
     "source_filename": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Source Filename"
     },
     "updated_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Updated At"
     }
    },
    "required": [
     "id",
     "slug",
     "display_name",
     "description",
     "active",
     "optional",
     "prompt_markdown"
    ],
    "title": "PerspectiveDetail",
    "type": "object"
   },
   "PerspectiveListResponse.15b0cf1": {
    "description": "``GET /api/perspectives/`` response body.",
    "items": {
     "$ref": "#/components/schemas/PerspectiveListResponse.15b0cf1.PerspectiveSummary"
    },
    "title": "PerspectiveListResponse",
    "type": "array"
   },
   "PerspectiveListResponse.15b0cf1.PerspectiveSummary": {
    "additionalProperties": false,
    "properties": {
     "active": {
      "title": "Active",
      "type": "boolean"
     },
     "description": {
      "title": "Description",
      "type": "string"
     },
     "display_name": {
      "title": "Display Name",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "optional": {
      "title": "Optional",
      "type": "boolean"
     },
     "slug": {
      "title": "Slug",
      "type": "string"
     },
     "source_filename": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Source Filename"
     },
     "updated_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Updated At"
     }
    },
    "required": [
     "id",
     "slug",
     "display_name",
     "description",
     "active",
     "optional"
    ],
    "title": "PerspectiveSummary",
    "type": "object"
   },
...
   "ValidationError.6a07bef": {
    "description": "Model of a validation error response.",
    "items": {
     "$ref": "#/components/schemas/ValidationError.6a07bef.ValidationErrorElement"
    },
    "title": "ValidationError",
    "type": "array"
   },
   "ValidationError.6a07bef.ValidationErrorElement": {
    "description": "Model of a validation error response element.",
    "properties": {
     "ctx": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "default": null,
      "title": "Error context"
     },
     "loc": {
      "items": {
       "type": "string"
      },
      "title": "Missing field name",
      "type": "array"
     },
     "msg": {
      "title": "Error message",
      "type": "string"
     },
     "type": {
      "title": "Error type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationErrorElement",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Lightroom Tagger Visualizer API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/perspectives/": {
   "get": {
    "description": "",
    "operationId": "get__api_perspectives_",
    "parameters": [],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PerspectiveListResponse.15b0cf1"
        }
       }
      },
      "description": "OK"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "List perspectives; optional ``active_only=true`` query.",
    "tags": [
     "perspectives"
    ]
   },
   "post": {
    "description": "",
    "operationId": "post__api_perspectives_",
    "parameters": [],
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PerspectiveDetail.15b0cf1"
        }
       }
      },
      "description": "Created"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "create_perspective_route <POST>",
    "tags": [
     "perspectives"
    ]
   }
  },
  "/api/perspectives/{slug}": {
   "delete": {
    "description": "",
    "operationId": "delete__api_perspectives_{slug}",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "slug",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "delete_perspective_route <DELETE>",
    "tags": [
     "perspectives"
    ]
   },
   "get": {
    "description": "",
    "operationId": "get__api_perspectives_{slug}",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "slug",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PerspectiveDetail.15b0cf1"
        }
       }
      },
      "description": "OK"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "get_perspective_route <GET>",
    "tags": [
     "perspectives"
    ]
   },
   "put": {
    "description": "",
    "operationId": "put__api_perspectives_{slug}",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "slug",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PerspectiveDetail.15b0cf1"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "update_perspective_route <PUT>",
    "tags": [
     "perspectives"
    ]
   }
  },
  "/api/perspectives/{slug}/reset-default": {
   "post": {
    "description": "",
    "operationId": "post__api_perspectives_{slug}_reset-default",
    "parameters": [
     {
      "description": "",
      "in": "path",
      "name": "slug",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PerspectiveDetail.15b0cf1"
        }
       }
      },
      "description": "OK"
     },
     "400": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Bad Request"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorBody.45d9b59"
        }
       }
      },
      "description": "Not Found"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ValidationError.6a07bef"
        }
       }
      },
      "description": "Unprocessable Content"
     }
    },
    "summary": "reset_perspective_default_route <POST>",
    "tags": [
     "perspectives"
    ]
   }
  },
...
 },
 "security": [],
...
}